恢复

+++ .github/Convert.sh
@@ -4,9 +4,14 @@ home="."
 
 Latest_version="$(curl "https://github.com/YAWAsau/backup_script/releases" -sL | awk -F "/YAWAsau/backup_script/releases/download" '{print $2}' | awk -F ".zip" '{print $1}')"
 Latest_version="$(echo ${Latest_version} | awk -F "/" '{print $2}')"
+echo "ReleaseVersion=${Latest_version}" >> ${GITHUB_ENV}
 
 [ ! -d "${home}/zip" ] && echo "- 创建zip目录" && mkdir -p ${home}/zip
 [ -f "${home}/zip/version" ] && Previous_version="$(cat ${home}/zip/version)" || Previous_version=""
+
+if [ "${Previous_version}" != "${Latest_version}" ]; then
+
+  echo "Judgement_Push=true" >> ${GITHUB_ENV}
 
   rm -rf ${home}/zip/v${Previous_version}.zip 1>/dev/null 2>&1
   echo "${Latest_version}" > "${home}/zip/version"
@@ -56,3 +61,4 @@ Latest_version="$(echo ${Latest_version} | awk -F "/" '{print $2}')"
 
   rm -rf ${home}/Download/
   rm -rf ${home}/v${Latest_version}/
+fi
